requirements.txt
setuptools~=69.0.2
wheel~=0.42.0
twine~=4.0.2
mypy~=1.7.1
black~=23.11.0
tox~=4.11.4
versioneer~=0.29
pyinstaller
pytest
pytest-mock

jsonschema
PyGithub
GitPython